\name{boss.generater}
\alias{boss.generater}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
boss.generater
}
\description{
  It is a Boss generater. 
}
\usage{
boss.generater()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	none
}
\details{
No arguments
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
No references.
}
\author{
Thomas Maltesen
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
dd

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
